Extract HyperliquidQuicknodeInfoClient for all QN routing
HyperliquidInfoClient is now public-HL only, delegates QN types.
QN_PROXIED_TYPES, post(), portfolio_state() live in the QN client.
